+++ zcash_client_backend/src/data_api.rs
@@ -1,12 +1,22 @@
 //! Interfaces for wallet data persistence & low-level wallet utilities.
 
 use std::{
-    collections::{BTreeMap, HashMap},
+    collections::HashMap,
     fmt::Debug,
     io,
     num::{NonZeroU32, TryFromIntError},
 };
 
+use self::scanning::ScanRange;
+use self::{chain::CommitmentTreeRoot, wallet::Account};
+use crate::{
+    address::UnifiedAddress,
+    decrypt::DecryptedOutput,
+    keys::{UnifiedAddressRequest, UnifiedFullViewingKey, UnifiedSpendingKey},
+    proto::service::TreeState,
+    wallet::{Note, NoteId, ReceivedNote, Recipient, WalletTransparentOutput, WalletTx},
+    ShieldedProtocol,
+};
 use incrementalmerkletree::{frontier::Frontier, Retention};
 use secrecy::SecretVec;
 use shardtree::{error::ShardTreeError, store::ShardStore, ShardTree};
@@ -18,20 +28,8 @@ use zcash_primitives::{
         components::amount::{Amount, BalanceError, NonNegativeAmount},
         Transaction, TxId,
     },
-    zip32::{AccountId, Scope},
+    zip32::Scope,
 };
-
-use crate::{
-    address::UnifiedAddress,
-    decrypt::DecryptedOutput,
-    keys::{UnifiedAddressRequest, UnifiedFullViewingKey, UnifiedSpendingKey},
-    proto::service::TreeState,
-    wallet::{Note, NoteId, ReceivedNote, Recipient, WalletTransparentOutput, WalletTx},
-    ShieldedProtocol,
-};
-
-use self::chain::CommitmentTreeRoot;
-use self::scanning::ScanRange;
 
 #[cfg(feature = "transparent-inputs")]
 use {
@@ -291,18 +289,18 @@ impl<T> Ratio<T> {
 /// this circumstance it is possible that a newly created transaction could conflict with a
 /// not-yet-mined transaction in the mempool.
 #[derive(Debug, Clone, PartialEq, Eq)]
-pub struct WalletSummary {
-    account_balances: BTreeMap<AccountId, AccountBalance>,
+pub struct WalletSummary<AccountId: Eq + std::hash::Hash> {
+    account_balances: HashMap<AccountId, AccountBalance>,
     chain_tip_height: BlockHeight,
     fully_scanned_height: BlockHeight,
     scan_progress: Option<Ratio<u64>>,
     next_sapling_subtree_index: u64,
 }
 
-impl WalletSummary {
+impl<AccountId: Eq + std::hash::Hash> WalletSummary<AccountId> {
     /// Constructs a new [`WalletSummary`] from its constituent parts.
     pub fn new(
-        account_balances: BTreeMap<AccountId, AccountBalance>,
+        account_balances: HashMap<AccountId, AccountBalance>,
         chain_tip_height: BlockHeight,
         fully_scanned_height: BlockHeight,
         scan_progress: Option<Ratio<u64>>,
@@ -318,7 +316,7 @@ impl WalletSummary {
     }
 
     /// Returns the balances of accounts in the wallet, keyed by account ID.
-    pub fn account_balances(&self) -> &BTreeMap<AccountId, AccountBalance> {
+    pub fn account_balances(&self) -> &HashMap<AccountId, AccountBalance> {
         &self.account_balances
     }
 
@@ -361,6 +359,9 @@ pub trait InputSource {
     /// The type of errors produced by a wallet backend.
     type Error;
 
+    /// The type used to track unique account identifiers.
+    type AccountId;
+
     /// Backend-specific note identifier.
     ///
     /// For example, this might be a database identifier type
@@ -384,7 +385,7 @@ pub trait InputSource {
     /// be included.
     fn select_spendable_notes(
         &self,
-        account: AccountId,
+        account: Self::AccountId,
         target_value: Amount,
         sources: &[ShieldedProtocol],
         anchor_height: BlockHeight,
@@ -424,6 +425,12 @@ pub trait InputSource {
 pub trait WalletRead {
     /// The type of errors that may be generated when querying a wallet data store.
     type Error;
+
+    /// The type used to track unique account identifiers.
+    type AccountId: Eq + std::hash::Hash;
+
+    /// Gets the parameters that went into creating an account (e.g. seed+index or uvk).
+    fn get_account(&self, account_id: &Self::AccountId) -> Result<Option<Account>, Self::Error>;
 
     /// Returns the height of the chain as known to the wallet as of the most recent call to
     /// [`WalletWrite::update_chain_tip`].
@@ -503,7 +510,7 @@ pub trait WalletRead {
 
     /// Returns the birthday height for the given account, or an error if the account is not known
     /// to the wallet.
-    fn get_account_birthday(&self, account: AccountId) -> Result<BlockHeight, Self::Error>;
+    fn get_account_birthday(&self, account: Self::AccountId) -> Result<BlockHeight, Self::Error>;
 
     /// Returns the most recently generated unified address for the specified account, if the
     /// account identifier specified refers to a valid account for this wallet.
@@ -512,26 +519,26 @@ pub trait WalletRead {
     /// account.
     fn get_current_address(
         &self,
-        account: AccountId,
+        account: Self::AccountId,
     ) -> Result<Option<UnifiedAddress>, Self::Error>;
 
     /// Returns all unified full viewing keys known to this wallet.
     fn get_unified_full_viewing_keys(
         &self,
-    ) -> Result<HashMap<AccountId, UnifiedFullViewingKey>, Self::Error>;
+    ) -> Result<HashMap<Self::AccountId, UnifiedFullViewingKey>, Self::Error>;
 
     /// Returns the account id corresponding to a given [`UnifiedFullViewingKey`], if any.
     fn get_account_for_ufvk(
         &self,
         ufvk: &UnifiedFullViewingKey,
-    ) -> Result<Option<AccountId>, Self::Error>;
+    ) -> Result<Option<Self::AccountId>, Self::Error>;
 
     /// Returns the wallet balances and sync status for an account given the specified minimum
     /// number of confirmations, or `Ok(None)` if the wallet has no balance data available.
     fn get_wallet_summary(
         &self,
         min_confirmations: u32,
-    ) -> Result<Option<WalletSummary>, Self::Error>;
+    ) -> Result<Option<WalletSummary<Self::AccountId>>, Self::Error>;
 
     /// Returns the memo for a note.
     ///
@@ -549,7 +556,7 @@ pub trait WalletRead {
     fn get_sapling_nullifiers(
         &self,
         query: NullifierQuery,
-    ) -> Result<Vec<(AccountId, sapling::Nullifier)>, Self::Error>;
+    ) -> Result<Vec<(Self::AccountId, sapling::Nullifier)>, Self::Error>;
 
     /// Returns the nullifiers for Orchard notes that the wallet is tracking, along with their
     /// associated account IDs, that are either unspent or have not yet been confirmed as spent (in
@@ -558,7 +565,7 @@ pub trait WalletRead {
     fn get_orchard_nullifiers(
         &self,
         query: NullifierQuery,
-    ) -> Result<Vec<(AccountId, orchard::note::Nullifier)>, Self::Error>;
+    ) -> Result<Vec<(Self::AccountId, orchard::note::Nullifier)>, Self::Error>;
 
     /// Returns the set of all transparent receivers associated with the given account.
     ///
@@ -568,7 +575,7 @@ pub trait WalletRead {
     #[cfg(feature = "transparent-inputs")]
     fn get_transparent_receivers(
         &self,
-        _account: AccountId,
+        _account: Self::AccountId,
     ) -> Result<HashMap<TransparentAddress, Option<TransparentAddressMetadata>>, Self::Error> {
         Ok(HashMap::new())
     }
@@ -578,14 +585,14 @@ pub trait WalletRead {
     #[cfg(feature = "transparent-inputs")]
     fn get_transparent_balances(
         &self,
-        _account: AccountId,
+        _account: Self::AccountId,
         _max_height: BlockHeight,
     ) -> Result<HashMap<TransparentAddress, Amount>, Self::Error> {
         Ok(HashMap::new())
     }
 
     /// Returns a vector with the IDs of all accounts known to this wallet.
-    fn get_account_ids(&self) -> Result<Vec<AccountId>, Self::Error>;
+    fn get_account_ids(&self) -> Result<Vec<Self::AccountId>, Self::Error>;
 }
 
 /// Metadata describing the sizes of the zcash note commitment trees as of a particular block.
@@ -700,23 +707,23 @@ pub struct ScannedBlockCommitments {
 /// decrypted and extracted from a [`CompactBlock`].
 ///
 /// [`CompactBlock`]: crate::proto::compact_formats::CompactBlock
-pub struct ScannedBlock<Nf, S> {
+pub struct ScannedBlock<Nf, S, A> {
     block_height: BlockHeight,
     block_hash: BlockHash,
     block_time: u32,
-    transactions: Vec<WalletTx<Nf, S>>,
+    transactions: Vec<WalletTx<Nf, S, A>>,
     sapling: ScannedBundles<sapling::Node, sapling::Nullifier>,
     #[cfg(feature = "orchard")]
     orchard: ScannedBundles<orchard::note::NoteCommitment, orchard::note::Nullifier>,
 }
 
-impl<Nf, S> ScannedBlock<Nf, S> {
+impl<Nf, S, A> ScannedBlock<Nf, S, A> {
     /// Constructs a new `ScannedBlock`
     pub(crate) fn from_parts(
         block_height: BlockHeight,
         block_hash: BlockHash,
         block_time: u32,
-        transactions: Vec<WalletTx<Nf, S>>,
+        transactions: Vec<WalletTx<Nf, S, A>>,
         sapling: ScannedBundles<sapling::Node, sapling::Nullifier>,
         #[cfg(feature = "orchard")] orchard: ScannedBundles<
             orchard::note::NoteCommitment,
@@ -750,7 +757,7 @@ impl<Nf, S> ScannedBlock<Nf, S> {
     }
 
     /// Returns the list of transactions from this block that are relevant to the wallet.
-    pub fn transactions(&self) -> &[WalletTx<Nf, S>] {
+    pub fn transactions(&self) -> &[WalletTx<Nf, S, A>] {
         &self.transactions
     }
 
@@ -794,9 +801,9 @@ impl<Nf, S> ScannedBlock<Nf, S> {
 ///
 /// The purpose of this struct is to permit atomic updates of the
 /// wallet database when transactions are successfully decrypted.
-pub struct DecryptedTransaction<'a> {
+pub struct DecryptedTransaction<'a, AccountId> {
     pub tx: &'a Transaction,
-    pub sapling_outputs: &'a Vec<DecryptedOutput<sapling::Note>>,
+    pub sapling_outputs: &'a Vec<DecryptedOutput<sapling::Note, AccountId>>,
 }
 
 /// A transaction that was constructed and sent by the wallet.
@@ -804,11 +811,11 @@ pub struct DecryptedTransaction<'a> {
 /// The purpose of this struct is to permit atomic updates of the
 /// wallet database when transactions are created and submitted
 /// to the network.
-pub struct SentTransaction<'a> {
+pub struct SentTransaction<'a, AccountId> {
     pub tx: &'a Transaction,
     pub created: time::OffsetDateTime,
     pub account: AccountId,
-    pub outputs: Vec<SentTransactionOutput>,
+    pub outputs: Vec<SentTransactionOutput<AccountId>>,
     pub fee_amount: Amount,
     #[cfg(feature = "transparent-inputs")]
     pub utxos_spent: Vec<OutPoint>,
@@ -817,14 +824,14 @@ pub struct SentTransaction<'a> {
 /// An output of a transaction generated by the wallet.
 ///
 /// This type is capable of representing both shielded and transparent outputs.
-pub struct SentTransactionOutput {
+pub struct SentTransactionOutput<AccountId> {
     output_index: usize,
-    recipient: Recipient<Note>,
+    recipient: Recipient<AccountId, Note>,
     value: NonNegativeAmount,
     memo: Option<MemoBytes>,
 }
 
-impl SentTransactionOutput {
+impl<AccountId> SentTransactionOutput<AccountId> {
     /// Constructs a new [`SentTransactionOutput`] from its constituent parts.
     ///
     /// ### Fields:
@@ -836,7 +843,7 @@ impl SentTransactionOutput {
     /// * `memo` - the memo that was sent with this output
     pub fn from_parts(
         output_index: usize,
-        recipient: Recipient<Note>,
+        recipient: Recipient<AccountId, Note>,
         value: NonNegativeAmount,
         memo: Option<MemoBytes>,
     ) -> Self {
@@ -859,7 +866,7 @@ impl SentTransactionOutput {
     }
     /// Returns the recipient address of the transaction, or the account id and
     /// resulting note for wallet-internal outputs.
-    pub fn recipient(&self) -> &Recipient<Note> {
+    pub fn recipient(&self) -> &Recipient<AccountId, Note> {
         &self.recipient
     }
     /// Returns the value of the newly created output.
@@ -999,6 +1006,7 @@ pub trait WalletWrite: WalletRead {
     ///
     /// Returns the account identifier for the newly-created wallet database entry, along with the
     /// associated [`UnifiedSpendingKey`].
+    /// Note that the unique account identifier should *not* be assumed equivalent to the ZIP-32 account index.
     ///
     /// If `birthday.height()` is below the current chain tip, this operation will
     /// trigger a re-scan of the blocks at and above the provided height. The birthday height is
@@ -1024,7 +1032,7 @@ pub trait WalletWrite: WalletRead {
         &mut self,
         seed: &SecretVec<u8>,
         birthday: AccountBirthday,
-    ) -> Result<(AccountId, UnifiedSpendingKey), Self::Error>;
+    ) -> Result<(Self::AccountId, UnifiedSpendingKey), Self::Error>;
 
     /// Generates and persists the next available diversified address, given the current
     /// addresses known to the wallet.
@@ -1033,7 +1041,7 @@ pub trait WalletWrite: WalletRead {
     /// account.
     fn get_next_available_address(
         &mut self,
-        account: AccountId,
+        account: Self::AccountId,
         request: UnifiedAddressRequest,
     ) -> Result<Option<UnifiedAddress>, Self::Error>;
 
@@ -1044,7 +1052,7 @@ pub trait WalletWrite: WalletRead {
     /// `blocks` must be sequential, in order of increasing block height
     fn put_blocks(
         &mut self,
-        blocks: Vec<ScannedBlock<sapling::Nullifier, Scope>>,
+        blocks: Vec<ScannedBlock<sapling::Nullifier, Scope, Self::AccountId>>,
     ) -> Result<(), Self::Error>;
 
     /// Updates the wallet's view of the blockchain.
@@ -1057,11 +1065,17 @@ pub trait WalletWrite: WalletRead {
     fn update_chain_tip(&mut self, tip_height: BlockHeight) -> Result<(), Self::Error>;
 
     /// Caches a decrypted transaction in the persistent wallet store.
-    fn store_decrypted_tx(&mut self, received_tx: DecryptedTransaction) -> Result<(), Self::Error>;
+    fn store_decrypted_tx(
+        &mut self,
+        received_tx: DecryptedTransaction<Self::AccountId>,
+    ) -> Result<(), Self::Error>;
 
     /// Saves information about a transaction that was constructed and sent by the wallet to the
     /// persistent wallet store.
-    fn store_sent_tx(&mut self, sent_tx: &SentTransaction) -> Result<(), Self::Error>;
+    fn store_sent_tx(
+        &mut self,
+        sent_tx: &SentTransaction<Self::AccountId>,
+    ) -> Result<(), Self::Error>;
 
     /// Truncates the wallet database to the specified height.
     ///
@@ -1167,7 +1181,7 @@ pub mod testing {
         consensus::{BlockHeight, Network},
         memo::Memo,
         transaction::{components::Amount, Transaction, TxId},
-        zip32::{AccountId, Scope},
+        zip32::Scope,
     };
 
     use crate::{
@@ -1218,6 +1232,7 @@ pub mod testing {
     impl InputSource for MockWalletDb {
         type Error = ();
         type NoteRef = u32;
+        type AccountId = u32;
 
         fn get_spendable_note(
             &self,
@@ -1230,7 +1245,7 @@ pub mod testing {
 
         fn select_spendable_notes(
             &self,
-            _account: AccountId,
+            _account: Self::AccountId,
             _target_value: Amount,
             _sources: &[ShieldedProtocol],
             _anchor_height: BlockHeight,
@@ -1242,6 +1257,14 @@ pub mod testing {
 
     impl WalletRead for MockWalletDb {
         type Error = ();
+        type AccountId = u32;
+
+        fn get_account(
+            &self,
+            _account_id: &Self::AccountId,
+        ) -> Result<Option<super::Account>, Self::Error> {
+            Ok(None)
+        }
 
         fn chain_height(&self) -> Result<Option<BlockHeight>, Self::Error> {
             Ok(None)
@@ -1296,34 +1319,37 @@ pub mod testing {
             Ok(None)
         }
 
-        fn get_account_birthday(&self, _account: AccountId) -> Result<BlockHeight, Self::Error> {
+        fn get_account_birthday(
+            &self,
+            _account: Self::AccountId,
+        ) -> Result<BlockHeight, Self::Error> {
             Err(())
         }
 
         fn get_current_address(
             &self,
-            _account: AccountId,
+            _account: Self::AccountId,
         ) -> Result<Option<UnifiedAddress>, Self::Error> {
             Ok(None)
         }
 
         fn get_unified_full_viewing_keys(
             &self,
-        ) -> Result<HashMap<AccountId, UnifiedFullViewingKey>, Self::Error> {
+        ) -> Result<HashMap<Self::AccountId, UnifiedFullViewingKey>, Self::Error> {
             Ok(HashMap::new())
         }
 
         fn get_account_for_ufvk(
             &self,
             _ufvk: &UnifiedFullViewingKey,
-        ) -> Result<Option<AccountId>, Self::Error> {
+        ) -> Result<Option<Self::AccountId>, Self::Error> {
             Ok(None)
         }
 
         fn get_wallet_summary(
             &self,
             _min_confirmations: u32,
-        ) -> Result<Option<WalletSummary>, Self::Error> {
+        ) -> Result<Option<WalletSummary<Self::AccountId>>, Self::Error> {
             Ok(None)
         }
 
@@ -1338,7 +1364,7 @@ pub mod testing {
         fn get_sapling_nullifiers(
             &self,
             _query: NullifierQuery,
-        ) -> Result<Vec<(AccountId, sapling::Nullifier)>, Self::Error> {
+        ) -> Result<Vec<(Self::AccountId, sapling::Nullifier)>, Self::Error> {
             Ok(Vec::new())
         }
 
@@ -1346,14 +1372,14 @@ pub mod testing {
         fn get_orchard_nullifiers(
             &self,
             _query: NullifierQuery,
-        ) -> Result<Vec<(AccountId, orchard::note::Nullifier)>, Self::Error> {
+        ) -> Result<Vec<(Self::AccountId, orchard::note::Nullifier)>, Self::Error> {
             Ok(Vec::new())
         }
 
         #[cfg(feature = "transparent-inputs")]
         fn get_transparent_receivers(
             &self,
-            _account: AccountId,
+            _account: Self::AccountId,
         ) -> Result<HashMap<TransparentAddress, Option<TransparentAddressMetadata>>, Self::Error>
         {
             Ok(HashMap::new())
@@ -1362,13 +1388,13 @@ pub mod testing {
         #[cfg(feature = "transparent-inputs")]
         fn get_transparent_balances(
             &self,
-            _account: AccountId,
+            _account: Self::AccountId,
             _max_height: BlockHeight,
         ) -> Result<HashMap<TransparentAddress, Amount>, Self::Error> {
             Ok(HashMap::new())
         }
 
-        fn get_account_ids(&self) -> Result<Vec<AccountId>, Self::Error> {
+        fn get_account_ids(&self) -> Result<Vec<Self::AccountId>, Self::Error> {
             Ok(Vec::new())
         }
     }
@@ -1380,16 +1406,16 @@ pub mod testing {
             &mut self,
             seed: &SecretVec<u8>,
             _birthday: AccountBirthday,
-        ) -> Result<(AccountId, UnifiedSpendingKey), Self::Error> {
-            let account = AccountId::ZERO;
+        ) -> Result<(Self::AccountId, UnifiedSpendingKey), Self::Error> {
+            let account = zip32::AccountId::ZERO;
             UnifiedSpendingKey::from_seed(&self.network, seed.expose_secret(), account)
-                .map(|k| (account, k))
+                .map(|k| (u32::from(account), k))
                 .map_err(|_| ())
         }
 
         fn get_next_available_address(
             &mut self,
-            _account: AccountId,
+            _account: Self::AccountId,
             _request: UnifiedAddressRequest,
         ) -> Result<Option<UnifiedAddress>, Self::Error> {
             Ok(None)
@@ -1398,7 +1424,7 @@ pub mod testing {
         #[allow(clippy::type_complexity)]
         fn put_blocks(
             &mut self,
-            _blocks: Vec<ScannedBlock<sapling::Nullifier, Scope>>,
+            _blocks: Vec<ScannedBlock<sapling::Nullifier, Scope, Self::AccountId>>,
         ) -> Result<(), Self::Error> {
             Ok(())
         }
@@ -1409,12 +1435,15 @@ pub mod testing {
 
         fn store_decrypted_tx(
             &mut self,
-            _received_tx: DecryptedTransaction,
+            _received_tx: DecryptedTransaction<Self::AccountId>,
         ) -> Result<(), Self::Error> {
             Ok(())
         }
 
-        fn store_sent_tx(&mut self, _sent_tx: &SentTransaction) -> Result<(), Self::Error> {
+        fn store_sent_tx(
+            &mut self,
+            _sent_tx: &SentTransaction<Self::AccountId>,
+        ) -> Result<(), Self::Error> {
             Ok(())
         }