+++ zcash_client_backend/src/data_api/wallet/input_selection.rs
@@ -17,7 +17,6 @@ use zcash_primitives::{
         },
         fees::FeeRule,
     },
-    zip32::AccountId,
 };
 
 use crate::{
@@ -149,7 +148,7 @@ pub trait InputSelector {
         wallet_db: &Self::InputSource,
         target_height: BlockHeight,
         anchor_height: BlockHeight,
-        account: AccountId,
+        account: <Self::InputSource as InputSource>::AccountId,
         transaction_request: TransactionRequest,
     ) -> Result<
         Proposal<Self::FeeRule, <Self::InputSource as InputSource>::NoteRef>,
@@ -315,6 +314,7 @@ impl<DbT, ChangeT: ChangeStrategy> GreedyInputSelector<DbT, ChangeT> {
 impl<DbT, ChangeT> InputSelector for GreedyInputSelector<DbT, ChangeT>
 where
     DbT: InputSource,
+    <DbT as InputSource>::AccountId: Clone,
     ChangeT: ChangeStrategy,
     ChangeT::FeeRule: Clone,
 {
@@ -329,7 +329,7 @@ where
         wallet_db: &Self::InputSource,
         target_height: BlockHeight,
         anchor_height: BlockHeight,
-        account: AccountId,
+        account: <DbT as InputSource>::AccountId,
         transaction_request: TransactionRequest,
     ) -> Result<
         Proposal<Self::FeeRule, DbT::NoteRef>,